writing to file and stdout
All i did was have two print statements to both the file and the
screen. I had to add a file pointer parameter to the allPaths method so
it could constantly be passed between paths. For some reason it repeats
now, and I couldn’t really figure out why. Maybe if we concatenate to a
char pointer, and then printed the entire char pointer at the end to
the file and to the screen, it might avoid duplicates?